split types from query
