+++ examples/github/__init__.py
@@ -1,1 +1,2 @@
 from .core import *  # noqa
+from .types import *  # noqa
